Split workspace packages and add independent releases
