+++ .github/workflows/cd.yml
@@ -1,0 +1,85 @@
+name: Publish package
+
+on:
+  release:
+    types: [published]
+  workflow_dispatch:
+    inputs:
+      package:
+        description: "Workspace package to build (for example url-to-markdown)"
+        required: true
+        type: choice
+        options:
+          - lobstergram
+          - md-to-telegraph
+          - url-to-markdown
+      version:
+        description: "Version already written in that package's pyproject.toml"
+        required: true
+        type: string
+
+permissions:
+  contents: read
+  id-token: write
+
+jobs:
+  publish:
+    name: Publish selected package
+    runs-on: ubuntu-latest
+    if: github.event_name == 'release' || github.event_name == 'workflow_dispatch'
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+
+      - name: Set up uv
+        uses: astral-sh/setup-uv@v7
+
+      - name: Select package
+        id: package
+        env:
+          RELEASE_TAG: ${{ github.event.release.tag_name }}
+          INPUT_PACKAGE: ${{ inputs.package }}
+        run: |
+          if [ "${{ github.event_name }}" = "workflow_dispatch" ]; then
+            package="$INPUT_PACKAGE"
+          else
+            case "$RELEASE_TAG" in
+              lobstergram-v*|md-to-telegraph-v*|url-to-markdown-v*)
+                package="${RELEASE_TAG%-v*}"
+                ;;
+              *)
+                echo "Release tag must be <package>-v<version>"
+                exit 1
+                ;;
+            esac
+          fi
+          echo "name=$package" >> "$GITHUB_OUTPUT"
+
+      - name: Verify package version
+        env:
+          PACKAGE: ${{ steps.package.outputs.name }}
+          RELEASE_TAG: ${{ github.event.release.tag_name }}
+          INPUT_VERSION: ${{ inputs.version }}
+        run: |
+          version="$(uv version --package "$PACKAGE" --short)"
+          if [ "${{ github.event_name }}" = "release" ]; then
+            expected="${RELEASE_TAG#*-v}"
+          else
+            expected="$INPUT_VERSION"
+          fi
+          test "$version" = "$expected" || {
+            echo "pyproject version ($version) does not match release version ($expected)"
+            exit 1
+          }
+
+      - name: Build package
+        env:
+          PACKAGE: ${{ steps.package.outputs.name }}
+        run: uv build --package "$PACKAGE" --out-dir "dist/$PACKAGE"
+
+      - name: Publish to PyPI
+        env:
+          PACKAGE: ${{ steps.package.outputs.name }}
+        run: uv publish "dist/$PACKAGE"/*
