+++ packages/md-to-telegraph/pyproject.toml
@@ -1,0 +1,21 @@
+[project]
+name = "md-to-telegraph"
+version = "0.1.0"
+description = "Convert Markdown to Telegraph DOM nodes"
+readme = "README.md"
+authors = [{ name = "Martín Gaitán", email = "gaitan@gmail.com" }]
+license = { text = "MIT" }
+requires-python = ">=3.12"
+keywords = ["markdown", "telegraph", "dom"]
+dependencies = [
+    "mistletoe>=1.5.1",
+]
+
+[project.urls]
+Homepage = "https://github.com/mgaitan/lobstersgram/tree/main/packages/md-to-telegraph"
+Repository = "https://github.com/mgaitan/lobstersgram"
+Issues = "https://github.com/mgaitan/lobstersgram/issues"
+
+[build-system]
+requires = ["uv_build>=0.7,<0.12"]
+build-backend = "uv_build"
