+++ packages/url-to-markdown/pyproject.toml
@@ -1,0 +1,24 @@
+[project]
+name = "url-to-markdown"
+version = "0.1.0"
+description = "Extract web pages and supported special URLs as Markdown"
+readme = "README.md"
+authors = [{ name = "Martín Gaitán", email = "gaitan@gmail.com" }]
+license = { text = "MIT" }
+requires-python = ">=3.12"
+keywords = ["html", "markdown", "readability", "github", "arxiv"]
+dependencies = [
+    "beautifulsoup4>=4.14.3",
+    "markdownify>=1.2.2",
+    "readability-lxml>=0.8.4.1",
+    "requests>=2.32.5",
+]
+
+[project.urls]
+Homepage = "https://github.com/mgaitan/lobstersgram/tree/main/packages/url-to-markdown"
+Repository = "https://github.com/mgaitan/lobstersgram"
+Issues = "https://github.com/mgaitan/lobstersgram/issues"
+
+[build-system]
+requires = ["uv_build>=0.7,<0.12"]
+build-backend = "uv_build"
